flask_api/requirements.txt
async-timeout==4.0.2
azure-core==1.28.0
azure-storage-blob==12.17.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
flask-serialize==2.1.3
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.7
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
permissive-dict==1.0.4
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
